Lowered Chimera to clear X Carriage
Lowered the Chimera mount to provide more clearance for the heater
block. Previously the heater block was close enough to melt the X
Carriage unless turned. This causes an issue with more surface area for
the fan to hit, which in turn causes the fan to cool the heater more
than it should.