Universal config, Titan mount accessory holes, duct work.
Added option for using a universal hot end mount height. This is
default as well. Allows for replacement of accessories instead of
carriage when swapping hotends. Option to use optimized height remains
if desired.
Added a runner to the fan duct to better direct airflow.
Added Accessory Mount holes to the E3D Titan Mount.
Cleaned up code a bit.